Adding some more logic to canvas cases.

+++ VeloceCRM.Entity/CanvasCase.cs
@@ -11,8 +11,8 @@ namespace VeloceCRM.Entity
         public string Name { get; set; } = "";
         [MaxLength(1024)]
         public string Description { get; set; } = "";
-        public int Starts { get; set; }
-        public int? Ends { get; set; }
+        public long Starts { get; set; }
+        public long? Ends { get; set; }
         public long CaseMasterId { get; set; }
         public int ActivityType { get; set; }
     }
